1. Use sysconf() to learn max number of open files on system instead of _POSIX_OPEN_MAX.  The number _POSIX_OPEN_MAX is simply the smallest number of open files allowed by the POSIX standard, but we need to record open files up the to the maximum number a user application may have on the current system.
2. Chang PreOpen() from taking only the file descriptor number as
an argument to taking also booleans specifying whether the file is
readable and/or writeable.  Allows PreOpen to deal with other files
than 0, 1, and 2 - needed to "pre-open" remote syscall and logging
sockets.

3. Change printf() to dprintf() so we get debugging output in the
shadow's log allowing running of programs which will want to redirect
stderr to a file belonging to the user.

4. Change all remote syscall numbers from SYS_XXX to CONDOR_XXX.  Now
the "SYS" values are used only for the local system calls, and
the "CONDOR" values are used consitently for all the remote stuff.

5. Supply _open() routine for OFS/1 which uses "pseudo" system call
_open() for routines like fopen().  The _open routine is just an
alternate name for open().

6. Make isatty() an undefined symbol so that it will get satisfied
by our custom routine, and not the system's routine which generally
calls ioctl() (ioctl is not supported for condor programs).