1. New rule for non_supported system calls.  The call is switched and sent to the shadow, but the receiver calls the CONDOR_NotSupported() rouinte instead of invoking the system call at the remote end.  Thing having to do with memory mapped files, signals, etc. fall into this category.
2. Add support for "pseudo" system calls.  These are special condor
functions carried out by the shadow, and can be invoked with the
REMOTE_sycall() routine.  The don't have switches since there is no
"local" version to invoke, but they do have senders and receivers.  The
senders are as usual, but the receivers must call hand written code in
the shadow, cause these are functions not supplied in the normal C
library.  Things like CONDOR_send_file and CONDOR_get_file fall into
this category.

3. Add support for extracted system calls.  These must be extracted
from the C (and renamed).  They are system calls which cannot be
sucessfully invoded by the syscall() routine.