Fixed the bug of condor_rm for PVM jobs.
A PVM universe job may not have a match for shadow record (why?).
In the function abort_job_myself(), instead of accessing srec->match,
we use IsSchedulerUniverse() to test if the job is a scheduler
universe job.  Also, in the dprintf statement,  we cannot dereference
srec->match if it is NULL.