This debugging module provides a simple mechanism of conditionally printing a message to the screen depending on the output level specified by the user at run time.
For example, if the debug level is set to 3 (VERBOSE), then messages above 3
will not print.  So a level 4 (DEBUG) message won't print, because the user
specified only messages up to 3 to print.  A level 3 (VERBOSE) or 2 (NORMAL)
would print, being less than or equal to 3.

This module allows the output level of the program to dynamically change without the need to recompile.  The small expense is that all debug messages are compiled in whether or not they are used during a given run.  I believe this is a small price to pay, for the convenience of dynamic adjustment of output level.