Configure size of SocketCache using NEGOTIATOR_SOCKET_CACHE_SIZE.
Implemented "preempt for rank".  Now, instead of always preferring to
match jobs with idle machines, the negotiator follows the job's rank
expression, even when it ranks an allocated machine over an idle
machine.  PREEMPT_REQUIREMENTS is still used to determine if a match
is preemptible, and PREEMPT_RANK is used to rank matches which the job
ranks the same.  Matches which don't require preemption get a
PREEMPT_RANK of FLOAT_MAX, so we always choose an idle machine when
the job doesn't care one way or the other.

Added D_BANDWIDTH dprintf messages, so we can keep a persistent log of
bandwidth allocation statistics (using the dprintf functionality added
previously to use different files for different debug levels).