+ changed _EXCEPT_File to __FILE__ instead of _FileName_.  The result is everything still works (i hope!) _without_ needing the    static char *_FileName_ = __FILE__ statement at the top of every source file using EXCEPT.  Why did we not do this YEARS ago????
+ For Win32: assert() is made the same as ASSERT(), which does an EXCEPT
instead of whatever the C runtime library does.  This way we (a) see a
message in our logs, and (b) exit with JOB_EXCEPTION return code, so our
parent process knows there was trouble.  We do not do this change on Unix,
since this same code is used by the user job in condor_syscall_lib, and
there we must really assert so we die with sigabort.