Some sneaky bug fixes:
1) On Linux, fixed a bug where the process's "start time" wasn't as
large as it should've been (it was an OS error) resulting in an age of
-1.  When we see an age < 0, we make the age = 0.

2) We could accidentally re-use dead pid entries in the hashtable,
resulting in ugly negative cpu usage times.  We now also check the
starting time of the entry we get back from the hashtable to make
sure that the pid is really who we think it is.

Also:

1) Cleaned up code that does the hashtable work - it's now in one
function instead of replicated 4 times.

2) Left detection in place for negative usage problems, just in
case they happen.  (We print an AAAAG!, some info, and drop core).