+ Fixed so we only care about qmgmt socket authentication results _if_ we call OwnerCheck(), and we only call OwnerCheck() on qmgmt commands which write to the queue.  Thus, read-only qmgmt commands, like condor_q -global, will work once again even if the only authentication method permitted is the local FILESYSTEM check.
+ Got rid of CheckConnection().  It was bogus.

+ Changed OwnerCheck so if the user is NULL, OwnerCheck fails.  A NULL user
nowadays means the relisock authenticate failed.  We do not have to worry
about "internal" schedd commands because we are carefull to only call
OwnerCheck if Q_SOCK is NULL; Q_SOCK is NULL if we are being called
internally from within the schedd.

+ Fixed some bugs dealing with resetting the state of some global/static
variables between connections, such as active_cluster_num and
active_owner_uid.

+ Got rid of some harmless (but annoying) compiler warnings.

+ Made some "char *" params type "const char *" where needed.