I found all instances of:
ctime((time_t*)&foo)

where foo is an int and replaced it with:

time_t arch_time = 0;
arch_time = foo;
ctime(&arch_time);

This is because on an alpha linux machine time_t is 8 bytes while int
is only 4 bytes. So typecasting the address of an int to a pointer that
will dereference 8 bytes will cause 4 extra bytes to be used in whatever
computation occurs. This fixes the behaviour of ctime returning a NULL
on too large a time.

Search for arch_time to see the relevent changes in the code.