Here lies cached versions of information you obtain from param(). If for any reason you need to put some code into the sysapi, and that code calls param(), place an initializer in this file so you may cache the variable. Some of these functions in sysapi get called a lot, so it is wise to cache the results.
The important calls in here are sysapi_reconfig() and
sysapi_internal_reconfig(). The former is called anytime someone changes
parameters in code the _uses_ the library. The latter is called _first_
in each sysapi_* and sysapi_*_raw function placed into the sysapi.

The _internal_ reconfig sees if the library has set up its tiny cache,
and of not does so. If it is called more than once, it reverts to
a no-op.

sysapi_reconfig() will ALWAYS reset up the cached variables with param.