+ changes to support classad "chaining" in the schedd (cluster ads).
+ if InitializeConnection fails, no need to set_user_id to nobody.

+ elminated lots of classad lookups on cluster & proc by using
ClassAdCollection method which returns both the ad _and_ the key, and
then just getting the cluster & proc out of the key.

+ commented out some unused functions, like DestroyClusterByConstraint

+ centralized the logic to cleanup a cluster into new function
called ClusterCleanup().

+ removed code to insert ATTR_SERVER_TIME into the ad; this is now
handled more efficiently in the AttrList:put() method

+ removed lookups for ATTR_CLUSTER_ID in order to determine if this
ad is the Header ad or not -- instead we just see if the key starts
with the character '0' (zero).  This is much faster and also detects
cluster ads.