For the argument F_FREESP in fcntl, it used to be that if start, len, and whence were all zero, it would be labeled an ftruncate(fd, 0).
However, a program was found that used fcntl F_FREESP with
whence == 0, len == 0, and start >= 0 to mean ftruncate the file to
start bytes long. I read the man page and sure enough, that is valid.
So I fixed it that if whence==0, len ==0, and start >= 0, to truncate
the file to start bytes long.