+ Completed implementation of _condor_vfprintf_va().
Supported things:
%u, %d, %o, %x, %i, %c, %s, %p, %%, all escape characters except \ooo and \x.

Unsupported things:
%f, %X, %e, %E, %g, %G, the L and l modifiers, the position precision
character '*', precision numbers, left justification, and printing
leading zeros.

I read the spec in the K&R book about what happens if you give a wierd
conversion character, and it said "implementation defined". So I usually
just copy the offending character into the output buffer.

Please look in condor_includes/condor_debug.h to change how bug you wish
the output buffer to be.

+ Added implementation of _condor_itoa(). This functions takes an int
like quantity and a base, and converts it to an ascii representation.

There were no libc functions called in the creation of this code. This is
done to allow us to use this code ANYWHERE in checkpointing without regards
to whether or not libc is currently mapped in or not. This is very important
and must be preserved if the code gets modified for some reason.