+ added support for file permission/ownership checking.  this allows us on NT to check if the user has permission to read/write to a file before FileTransfer reads/writes.
+ RegisterCallback() method now takes a C++ method instead of just a
function pointer

+ added methods TotalBytesSent() and TotalBytesReceived()

+ added support to conditionally tranfer back "intermediate" files to
the server's SPOOL directory.  this is used by the starter to transfer
files at VACATE time (as opposed to just transfering back at job exit).
this transfer is ensured to be atomic by use of a temp directory and
a commit file.

+ transfer_<input|output>_files must use a comma to seperate multiple files.
this allows filenames with spaces to work.

+ the client side will set the file's modify time backwards 3 minutes.
this helps to ensure we get the right files when we later compare
modify times, even if the OS (aka WinNT) takes a while to sync the changes
to disk.

+ simplified the code by removing "static" from several methods, and
passing a "this" pointer to the ones which remained static.