+ we used to make certain we did not send a signal to any pid below 10. this was to restrictive.  changed so we do not signal any pid below 3.
+ complete overhaul of how we send a softkill on WinNT.  we now
find the window station and desktop where the job is running, and send
the job a WM_CLOSE.

+ all dprintfs dealing with sending softkills or hardkills now go to
D_PROCFAMILY instead of D_DAEMONCORE.

+ on NT, when we create a new thread, we initialize that thread with the
same priv_state as the parent thread.  then the priv_state is kept in
thread local storage (since user credentials are per thread on NT, unlike
Unix where it is per process).