+ added GetNextJobByCluster() method.  it is much more efficient than getting jobs by constraint with a constraint on the cluster number.
+ got rid of GetJobList.  the one place where we used it, it was not
neccesary, and just slowed things down.

+ FindRunnableJob() changed so it works like it did before _if_ the
negotiator did not send us the startd ad.  However, if we have the startd
ad, then FindRunnableJob will search the cluster for the highest prio
job which matches with the startd ad.  If we do not find another job in
the cluster, we search the entire queue for jobs submitted from the same
user for the highest prio job which matches with the startd.