+ support for storing the startd ad, sent to the schedd by the negotiator, in the match record.
+ changes to how we ask the negotiator to reschedule: the reschedule
request is now sent via UDP instead of TCP to get rid of deadlock
situations (i.e. if a condor_reschule or condor_submit happens while
we are negotiating, the schedd sends a reschedule to the negotiator which
does not repond because it is currently negotiating with us!).
In addition, the reschedule_request_pending stuff was removed.  This
was the wrong thing to do; sending the request as UDP was the right thing.

+ on WinNT, added some stubs so we do not have to link/port the
checkpoint server client library, and other fixes.