Two things.
Used D_CKPT and carved out a decent region of debug statements that will
be very verbose when you turn on that flag. Kept some of the D_ALWAYS
statments, mostly for critical failures, and places in the checkpointing
code where it would be good to have that knowledge all the time.

Changed it so that we block condor signals during the enitre restart
process. I realize that we went through great lengths to make it so that
we can die early if the owner comes back during a long restart, but it
didn't work all the way. The problem is that on IRIX, the signal trampoline
code is in libc, so if we mmap /dev/zero over libc, and get a signal,
we will blow up even before getting to our signal handler. :(

At a later date, I will revisit this and fix it.