fixed bugs involving spaces being removed from various macros in condor submit description files (like spaces between job arguments).
the problem getline() is used both for reading the condor_config file,
as well as the submit and dagman description files.  in condor_config, we
can safely throw away all whitespace, so getline() did an ltrunc() and all
was safe.  but whitespace in the job submit file could be critical, like
the whitespace between command-line args.  the quick patch was to
create getline_implementation() which takes the buffer size to use as
an argument.  when we read condor_config, we now call getline_implementation
and specify a small/efficient buffer of 128 bytes.  this way our daemons do
not leave a big buffer sitting in memory forever.  but when condor_submit
or condor_dagman call getline(), it calls getline_implementation with a
POSIX_ARG_MAX buffer size.