Whoo boy. Alright, I finally fixed the problem where daemons would send mail from superuser instead of comming from condor. There are some code changes that I made that affects NT and unix.
In the email_open function, I abstracted away how the mailerstream gets
opened. Now email_open_implementation() gets called just when the mailerstream
is about to be opened. The arguments passed to it are the mailer we wish to
use and the actual mailer command. Now based upon if you are NT or unix, you
will get completely differet code for email_open_implementation. The NT stuff
stayed the same as per the revision before this one, it is just in a more
coherent place.

However, the unix stuff is now all different. The shadow now forkes a child
that changes its real uid to the condor account, and sets up the environment
with condor goodies to try and get the mailer to send mail from the condor
account. Some interesting details with pipe and fdopen are done as to make the
code backwards compatible to how the original email_open worked(it returned
a FILE*).

Some oddities are that the priv state code must come before the pipe closing
and duping operations and I have to turn off logging for the child. I don't
know why the ordering of the code and the logging mattered, but it did across
the various OS'es. I wouldn't get mail from condor, or I wouldn't get mail at
all on some OS'es.