The "new" template instantiation convention for classads:   - instantiations for classads go in instantiation.C   - instantiations common to both collection client and server go in       coll-comm-inst.C   - instantiations unique to collection client go in coll-client-inst.C   - instantiations unique to collection server go in coll-serv-inst.C   - instantiations for experimental features go in exp-inst.C
The idea is to separate out instantiations into different object files such
that programs only requiring a subset of the functionality of the classad
library will only get the instantiations required for supporting that
functionality.  This reduces binary sizes by quite a bit.

The possible functionalities planned for in the above 5-way decomposition are:
  - classads only
  - classads and client collections
  - classads and server collections
  - classads and client and server collections
and all of the above with "experimental features"