Alrighty, I had slightly misunderstood Doug T. when we talked about making stat go through the file table.
Basically I now force stat(), fstat(), and lstat() through the file table.

The order of operations for those calls are:
user program traps into our system calls,
we go from whatever switch we end up in to the Filetable,
the file table set SYS_UNMAPPED and goes back into the switches,
the switches do the local or remote syscall as needed,
back in the filetable frame we fix up the sizes and mod/acc times of the file
	to be consistant with the logical file we have in the buffer,
we return the result back to the user.

Linux, and IRIX were annoying.... as always.