There is a problem on beak where people who are supposed to get email from standard jobs completing don't. Now I tried to debug this a long time and could never reproduce the problem on any other machine except beak. If you run a personal schedd(with the installed binaries), then things are ok and you always get the mail.
I set up condor as root on bluejay(a machine like beak) to no avail. I always
got the mail correctly, even if it did end up in the mailq because it is an
unsup machine, point is that it was there.

So since it looks like a permissions problem, I added a chdir("/") and a
umask(0) to the forked mailer process to try and alleviate this problem
and made the exception conditions a lot more verbose so when it happens
again maybe we'll get a clue why. But since beak is such a heavily used
production server, it would be very difficult of me to install the new
shadow for debugging purposes.

I currently don't know exactly what is wrong with beak not giving mail to
users whose standard jobs complete. But hopefully this fix will do something
useful concerning it.