A few changes made in the LINUX portions of this file:
+ In getProcInfo, I changed the junk types that are fscanf()'d in to
be longs instead of ints. This is because Alpha Linux can have 64-bit
values in the /proc entry that we are inspecting. Reading a 64 bit number
into a 32 bit space will smash the stack.

+ Fixed a memory leak in do_usage_sampling caused by some errant code that
Jim B. checked in. When determining if a procHashNode is "slack", the phn
was set to NULL instead of being removed from the hash table. I fixed it
so that a removal from the hash table occurrs.

In a related thing, it appears the the remove() method in the hash table
template does not do the right thing when removing a set of data with
the same key.  In addition to having an insert with the key and the data,
we should also have a remove with the key and the data. Right now there
is only a remove with the key(and it appears to just remove the data
associated with it in hash table order, which leaves dangling pointers).
This problem is not architecture related and can happen on any of our
platforms.

Because of the bug Jim B. introduced, this above case with remove began
to happen and the hash table code caused segfaults in this context because
a Lookup later wouldl return a dangling pointer.

+ When converting jiffies to seconds, you must divide by HZ instead of
a hand derived constant that looked right. :) HZ is radically different
between linux architectures. This fixed a bug where the birth time of a
process was miscalculated and caused errors in the statistics.
I fixed this problem in LINUX portions of this file as other OSes don't have
a jiffies measurement of time.

This next change affected all OSes.

+ In the constructor for procHashNode, instead of initializing the
"garbage" field to false, it was declared as a boolean local variable
that was then initialized to false, causing the true garbage field to be
just that after construction--garbage. :) This could, and most likely did,
play havoc with the procHashNode garbage collector in do_usage_sampling.