Added some attributes to begin codification of suspension events in the old shadow and old starter. Specifically, I added:
ATTR_TOTAL_SUSPENSIONS
ATTR_LAST_SUSPENSION_TIME
ATTR_CUMULATIVE_SUSPENSION_TIME

These attributes are kept in the "in memory" shadow job ad, but when the job
checkpoints or vacates, this information is updated to the job queue log
job ad. This means that if the jobs un/suspends numerous time and then the
shadow gets killed or the starter dies, that information is lost. Maybe later
it'll get fixed in a better way.