This is a big and dumb hack to force user job un/suspend events into the Condor system.
I added log_old_starter_shadow_suspend_event_hack() which takes
the strings that would have been written out to the shadowlog on the
CLIENT_LOG socket  and decides if a starter suspend event happens. It does
this by inspecting magic strings in the shadowlog. If it is determined
that such events have happened, then the events are recorded into the
user log and then the in memory job ad for the shadow.