This is such a bad hack.
Just AFTER suspending the user jobs, write out a magic string to the
CLIENT_LOG(which the starter isn't supposed to touch) that contains
information about suspending the job.

Then BEFORE unsuspending the jobs, write out info to the CLIENT_LOG
that you are about to unsuspend a user job.

This allows the shadow to parse the job's log socket and figure out if it
has been suspended or not for the purposes of creating user log entries and
keeping statistics about the information.

At every Checkpoint and Vacate, the information is stored persistantly
into the schedd. If anything happens to the starter which a job is
suspended (like it was killed), then the statistics information from
the last checkpoint/vacate is lost because the shadow only updates them
persistantly at that time.

I proilly should have written a loop around write(), but this code should
die anyway and I'll give it another reason.