+ Several fixes to get things to actually work in the general sense.
+ Fixed several silly bugs which resulted in the gridmanager dumping
core during typicaly operation.

+ Call globus_poll every second instead of every 5 seconds; every
five secconds was too slow since GASS transfers of the initial
executable, etc, was taking much too long.

+ call setUpdate() from inside of addJobUpdateEvent(), instead of
trying to remember to call it from all the callsites (which we
didn't do correctly).

+ added several dprintf messages

+ don't call job_probe right away when we first register the timer;
instead call it after 10 seconds (more likely there will be a job)

+ on startup if we see a job which has already been submitted to
Globus, we no longer just ignore it.  Instead we start "watching"
it again (place the contact string in our hashtable, setup to
get async events from Globus, etc).  This provides robustness
in the situation that the submit machine is restarted before
our Globus jobs complete.

+ When a job completes, actually remove it from the queue instead
of leaving it sitting there.

+ Re-ordering: we used to write events to the userlog/send email AFTER
removing a job from the queue.  Now changed so we write events/send email
BEFORE removing the job from the queue.  Now if an untimely crash occurs,
we may get an event written twice in the userlog, but that's better
than not having the event logged at all.

+ moved setUpdate() out from the GridManager class and into the global
namespace, since we need to call it from ::addJobUpdateEvent()