Added the ability to specify old shadow policies about what to do when a user job exhibits condition <x>:
These are expressions evaluated periodically in the shadow:

periodic_hold
	When this becomes true during execution, then put the job on hold.

periodic_remove
	When this becomes true during execution, remove the job from the queue.

on_exit_hold
	When this becomes true after a job exits, put it on hold

on_exit_remove
	When this becomes true after a job exits, then let it leave the queue,
	otherwise, keep it in the queue returning to Idle state.

Known problems:

The information condor gives back to you about what it did when these
policies occur is marginal at best. This MUST be fixed in a future release
or else we're gonna get a lot of rust. For example: If periodic_remove
becomes true at _any_ point(possibly even in the middle of a checkpoint),
the job will exit as if it had exited normally with status 0. Nowhere
is it mentioned that _Condor_ removed the job, not the job itself had
finished. This should be a message in the termination event, whether
or not this happened. Other information in the log files should be made
available as well. But this is the important one.