Added a couple things to support old shadow side policy management of what Condor should do with a job when the job experiences certain conditions.
When a job exits, determine if it exited by a signal or not and set
ATTR_ON_EXIT_BY_SIGNAL correctly(see condor_sttributes.h logfile).
Then set up the ATTR_ON_EXIT_SIGNAL or ATTR_ON_EXIT_CODE correctly to reflect
what the job either dies by, or what it returned.

If a condor_rm or condor_vacate is performed, then reset the user specified
hold/remove expressions(from the .cmd file) to defaults and let nature run
its course with the job. This is so you don't accidentily remove a job from
the queue when it was really supposed to be vacated and whatnot.

Also, I fixed up the HoldJob() function to dprintf a message about what it
is about to exit with so the ShadowLog looks more coherent. Otherwise the
ShadowLog looks like the shadow segfaulted or something because it seems to
stop in the middle of something and then a new shadow will start beneath it.

TODO:

	Add another parameter to log_termination() that will be a char *.
	This shall then be added to a modified JobTerminationEvent so that the
	code can tell the user that the job terminated on it own, or Condor
	forcibly terminated it. Otherwise, the user won't know that the job
	left the queue because Condor told it to leave the queue, or that it
	finished on its own.