Since Solaris 2.8 doesn't have a true utmp, and the implementation of STARTD_HAS_BAD_UTMP = True has _serious_ problem under OSes that don't use a real /dev, I added support for utmpx in utmp_pty_idle_time().
Under Solaris 2.8, we read in the utmpx file using the associated utmpx
structure just like we normally do for struct utmp on machines that use it.

The STARTD_HAS_BAD_UTMP bug is this:

While finding out all of the idle times, we make a new Directory object for
"/devices/pseudo" and iterate over each file looking for tty's. The bug comes
in that dev_idle_time assumes(as in strcats the filename after /dev and tries
to stat that filename) all filenames will be under "/dev". This means
that devices scanned under the Directory object's list of file names won't
exist in /dev. The fix is to propogate the directory name you want to look
for the device under into dev_idle_time().

On machines with a real /dev structure, this bug is not shown. Currently,
Thu Mar  8 16:35:36 CST 2001
this means only solaris 2.8.

However, since I added utmpx support, we can now happily not run the
STARTD_HAS_BAD_UTMP code that contains the bug.