+ Made GetMyType() method return a const char* instead of just a char*.
+ Changed GetMyType() method so it will distinguish between a SCHEDD_ADTYPE
and a SUBMITTER_ADTYPE.  This fixes problems in the v6.3 condor_negotiator
such as trying to negotiate for hostnames instead of usernames
and bad ScheddLimits resulting in jobs never running with a personal condor.
It was fixed this way (which is actually a yucky hack) in order to
preserve backwards compatibility with earlier Condor releases.  Actually
changing things so submittor ads are stored in the collector with
SUBMITTER_ADTYPE would break compatibility and/or flocking with older
pools.