Big checkin, sorry...
+ In the code which continues to negotiate with a schedd past the
ScheddLimit in order to handle startd rank properly:  changed so
we will only start a job due to startd rank if the startd is already
claimed.  This way we no longer need to assume that Rank > 0 means
the startd prefers a job.  So setting RANK=1 in a condor_config file
no longer will cause the first user's jobs to all run...  It now works
as you'd expect with respect to startd rank.

+ We now get all the submittor ads from the Collector (instead of
just the Submittor ads where num_idle_jobs > 0, like we used to do
in v6.2).  This is a good thing, since it will allow our pie slicing
code to compute the right values.  However, there is no need to actually
talk to a schedd which has no idle jobs - so this optimization was added.

+ Must initialize reasons for match failure before we examine them!

+ Changed so we no longer create a StartdAds list and a ClaimedStartdAds
list.  The ClaimedStartdAds list was just used by the Accountant to check
if matches are still active.  But the bug was this list did not include
Matched startds.  So if the schedd took more than one negotiation cycle
to claim, the Accountant would remove its knowledge of the match
because it assumed the machine was down because it did not see its
startd ad.  So now we just pass the Accountant the
StartAds list (and the accountant
already did its own checking to figure out the state).

+ Fixed major bug: The scheddAds list must be populated with *submittor*
ads, not actual schedd ads!

+ If we match a users job due to startd rank after the ScheddLimit has
expired (new feature to v6.3), now the code will consult
the PREEMPTION_RANK expression to help determine which startd to grab.