+ Fixed stat and lstat.
It turns out the FILESYSTEM_DOMAIN is _extremely_ liberal in the
directories it forces condor to think as being shared(when using USE_NFS
or USE_AFS is true). For example, it turns out the Condor thinks /tmp is
shared among machines with similar FileSystem_Domain attributes. Because
of this, in a standard universe job, it makes the user job open up files
in /tmp on the execute machine instead of the intuitive solution of the
submit machine.

Now, I didn't fix that problem, but it sure the hell does need fixing.

What was wrong with stat and lstat was that they were hard coded to
always fall back to remote io if the file they were looking for hadn't
already been entered into the file table. Falling back to remote io
(on a filesystem that is actually shared) is always correct, so I guess
that's why it was coded that way. However, since /tmp isn't actually
shared, this caused a serious problem. Stat would fail on files in /tmp
if they were not already in the filetable database.

So I fixed it so that if the file isn't in the filetable, then ask the
shadow where the location of the file is and then set the syscall flags
to reflect what type of call should happen(local or remote), and then
call ::stat or ::lstat. This follows the behavior of the rest of the system
calls in condor.