+ The ordering of some steps in checkpointing were incorrect. The steps as I found them were 1-save the memory image, 2-save the signal state of the program, 3-checkpoint the file table. The ordering _should_ have been 1-save the signal state of the program, 2-checkpoint the file table and 3-save the memory image of the code. I fixed it to be this.
However, while debugging this, I discovered that we do all sorts of
libc calls to malloc, new, delete, gethostbyname and other junk while in the
signal handler. This is a big nono and will cause us problems I'm sure.

We need to reconcile all of the shiny new features going into the standard
universe with the fact it is being called inside a signal handler. :)