Because we are increasingly using classads as internal data structures in Condor, we need some nice accessor functions to things we'd like to do with classads.
This file is supposed to hold the support functions we use with classads
to make them easy to use for this purpose.

Currently, the stuff in here is to do attribute dirty caching in a
classad.  Now when you update an attribute, you can set the attribute
as dirty in the classad itself, and later do something based upon that if you
find that attribute dirty later on.

You can do basic things like:
SetAttrDirty()
SetAttrClean()
IsAttrDirty()
AnyDirtyAttr()

This is good enough for most purposes. The header file should hold more
documentation eventually on the exact semantics of this cache.