Changed how the specific shadow intstance gets derived:
Now, a ShadowInitializer class gets created in the main_init() call which
performs the work of getting the job ad from the schedd(using daemoncore
callback and timers). Once this is done, the class inspects the job ad
to see what kind of universe it is, and then calls the appropriate new
with the right derived shadow type. Then it calls Shadow->init(...) with
the right arguments.

Although it isn't implemented, the code is there to allow the schedd to force
the jobad to the shadow with a UDP packet. If the shadow doesn't receive this
UDP packet in one second from start up, then it will connect back to the schedd
with good old ConnectQ and get teh job ad the old way. This preserves backwards
compatibility.