Gridmanager can now infer what events have been logged for a job based on its classad on a restart (of the gridmanager).
Rearranged the GlobusJob state machine a bit. Now, each state only has a
single "blocking" call (either to the schedd or to the GAHP library.
Besides simplifying the logic, this will ensure that different calls to the
GAHP library can't be mistaken for the same call (because we can force a
new call on every state change). Jobs becoming held or removed at random
points in the state machine is also handled better.