Added a completely new interface for the user job policy determination. Please see the big comment before the UserPolicy class at the end of the user_job_policy.h file for a good explanation of the interface.
The short version of it is that you pass a classad to a policy class and then
call AnalyzePolicy() on it. It will either return with STAYS_IN_QUEUE,
REMOVE_FROM_QUEU, HOLD_IN_QUEUE, or UNDEFINED_EVAL. The FiringExpression()
method will tell you expression fired to give you the above answer if it
isn't STAYS_IN_QUEUE.

This code isn't tested as of yet, but it franks and looks like it should work.
Derek needed it before I could test it in a real way, but the code is very
simple so if it blows up, it should be extremely easy to find.