Got rid of a lot of commented out old code.
Had to change stdRankCondition and preemptRankCondition to eval to true if
"CurrentRank" is undefined.  This seems to be the proper behavior since it
matches the behavior of V6.2 and V6.3 at this time.

Fixed some bugs that are related to bridging the gap between old ClassAds and
new ClassAds.  Basically, in new ClassAds you have to explicitly say
"other.attribute" to refer to an attribute the target classad.  This is not
the case in old ClassAds.  So there is a lot of hacky code in condor_q to try
to have it both ways.  When 6.4 is merged with the trunk and condor_q gets a
big rewrite this hacky stuff should go away and the MatchClassAd object should
be used instead.