Added code to bail out of the authentication with a failure if a CEDAR code returns a failure code. It used to be that if the shadows went away for some reason (most likely because the schedd is too slow) the schedd would continuously time out with 20 second timeouts trying to tell the shadows that they didn't authenticate correctly. :) Now as soon as the first cedar failure code returns(probably because of a closed socket), we just mark the authentication as failed and return.
I tried to take care with the clean up of memory and generated
files and such during this bailing procedure but the Rendevous
file name was interesting to deal with at this stage of the code(in
condor_auth_fs.C). See, in another branch there is code that deals with
how the rendevous directory memory is handled. So I'll leave it alone
for now, but after the merge, it should be looked at again.