This file used to be condor_sysapi/analyze_exec.c and you can look at that file to figure out what the previous logs were.
Not only was this file renamed to be a .C file to enable use of C++, but
for LINUX, the bfd support was ripped out because libbfd is a nightmare
to use across differing linux distributions. So, now under linux(only)
we use the version object to detect whether or not a standard universe
executable has been linked with Condor or not. The chances of a random
executable have the Condor version strings in it without being linked
with Condor are astronomical.