Fixed many bugs.   Behold:
+ when writing to the user log, don't just flush stdio buffers... actually
sync to disk before releasing the write lock.

+ when reading from the user log, for some reason we obtain a write
lock on the log file.  however, we open the file O_RDONLY (because,
after all, we do just read from it).  however, on RedHat 7.2, the
kernel will never block on a *write* lock if the file was opened
read-only.  thus the user log code attempted to read from the log
file at the same time when a potential record was being written.  the
current fix is to open the file O_RDWR so locking works on Linux.
however, i think soon we will change this write lock to a read lock...

+ method readEvent() used to block for one second before returning
ULOG_NO_EVENT.  now it returns right away if there is nothing to read.

+ if readEvent sees a "partial" event (normally should not happen, but
it tries to be robust), the code used to sleep for 1 second and try
again.  Changed so we first release the write lock before sleeping. :^).

+ if readEvent fails to read the event number, re-instantiate the
event object after we try to read it a second time.

+ if readEvent() sees a partial event and successfully reads it
a second time, we had a bug which would cause the *subsequent*
readEvent() to always fail.  Now fixed by synchronizing the log
before calling ULOG_OK after successfully reading the log on a retry.