Added the following config file parameters to work around apparent bugs in Globus 2.2 :
GRIDMANAGER_CONNECT_FAILURE_RETRY_COUNT -  If the gridmanager tries
to connect to a remote gatekeeper or jobmanager, and it fails to do
so, then automatically retry the connect X number of times as
specified by this entry.  Default value is 3.

GRIDMANAGER_CONNECT_FAILURE_RETRY_INTERVAL - Amount of time,
in seconds, to delay between connect retries.  Defaults to 5 secs.