Changes to support the creation of dot files that can help users visualize dags.
Dot is part of the graphviz package:

   http://www.research.att.com/sw/tools/graphviz/

Dot was made to draw dags, and we make dags, so it was a natural fit.
Besides, there have been two external projects where people tried to
use dot to draw dags, but didn't do it as well or as efficently as
we can do it from within DAGMan.